Merged in EC-1120_DelincenseAfterLeavingProcess (pull request #2494)
EC-1120 Delincense After Leaving Process
Added check on practitioner profile view for coach, this checks if the practitioner should be delicensed based on the last PQA or re-accreditation. Normally the coach is shown the de-license process at the end of the visit if needed, but they can leave early, so the profile is then a backup.

Approved-by: Liana Wiehahn